requirements.txt
langgraph
langchain
langchain-google-genai
boto3
python-dotenv
tldextract
beautifulsoup4
mcp
flask
flask-socketio
